# public/openapi.yaml
openapi: 3.0.3
info:
  title: Edusm User & Admin Management API
  version: 1.0.0
  description: API for creating users/admins, login, and role-based validation with token and superadmin_email.
servers:
  - url: https://next-js-online-learning-app.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local development server
paths:
...
  /buy/chapter:
    post:
      summary: Add or update or deactivate chapter
      operationId: addOrUpdateDeactiveChapter
      tags:
        - Admin - Buy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chapter_id
                - subject_id
                - user_id
                - purchase_date
                - validity
                - token
                - email
                - user_type
                - task_type
              properties:
                chapter_id:
                  type: integer
                  example: 1
                subject_id:
                  type: integer
                  example: 1
                user_id:
                  type: string
                  example: "user@example.com"
                purchase_date:
                  type: string
                  format: date
                  example: "2025-04-21 14:19:20"
                validity:
                  type: integer
                  example: 180
                token:
                  type: string
                  example: "abc123securetoken"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                user_type:
                  type: string
                  example: "super_admin"
                task_type:
                  type: string
                  example: ""
      responses:
        '200':
          description: Chapter purchase updated successfully
        '409':
          description: Chapter added successfully
        '400':
          description: Invalid user input
        '403':
          description: Token verification failed
        '500':
          description: Server error
...
  /buy/class:
    post:
      summary: Add or update or deactivate Class
      operationId: addOrUpdateDeactiveClass
      tags:
        - Admin - Buy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - class_id
                - user_id
                - purchase_date
                - validity
                - token
                - email
                - user_type
                - task_type
              properties:
                class_id:
                  type: integer
                  example: 1
                user_id:
                  type: string
                  example: "user@example.com"
                purchase_date:
                  type: string
                  format: date
                  example: "2025-04-21 14:19:20"
                validity:
                  type: integer
                  example: 180
                token:
                  type: string
                  example: "abc123securetoken"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                user_type:
                  type: string
                  example: "super_admin"
                task_type:
                  type: string
                  example: ""
      responses:
        '200':
          description: Subject updated successfully
        '409':
          description: Subject added successfully
        '400':
          description: Invalid user input
        '403':
          description: Token verification failed
        '500':
          description: Server error
...
  /blog:
    get:
      summary: Fetch blogs
      description: Fetch all blogs, a single blog by `blog_id`, limited blogs, or blogs by `cat_id`.
      operationId: getAllBlog
      tags:
        - Blog
      parameters:
        - in: query
          name: blog_id
          schema:
            type: string
          required: false
          description: ID of the blog to fetch a specific blog
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
          required: false
          description: Limit the number of blogs returned
        - in: query
          name: cat_id
          schema:
            type: string
          required: false
          description: Filter blogs by category ID
      responses:
        '200':
          description: Blogs fetched successfully
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Blog'
                  - $ref: '#/components/schemas/Blog'
        '404':
          description: Blog(s) not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        components:
          schemas:
            Blog:
              type: object
              properties:
                blog_id:
                  type: integer
                cat_id:
                  type: integer
                title:
                  type: string
                content:
                  type: string
                created_at:
                  type: string
                  format: date-time
            Error:
              type: object
              properties:
                error:
                  type: string
                details:
                  type: string
    
    post:
      summary: Add a new blog
      operationId: addBlog
      tags:
        - Blog
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                cat_id:
                  type: integer
                  example: 2
                title:
                  type: string
                  example: "New Blog Post"
                slug:
                  type: string
                  example: "new-blog-post"
                description1:
                  type: string
                  example: "Main content of the blog."
                description2:
                  type: string
                  example: "Additional content of the blog."
                create_date:
                  type: string
                  format: date
                  example: "2023-04-21"
                image1:
                  type: string
                  format: binary
      responses:
        '201':
          description: Blog added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Blog added"
                  id:
                    type: integer
                    example: 1
        '400':
          description: Invalid input data
        '500':
          description: Internal server error
    
    put:
      summary: Update a blog
      operationId: updateBlog
      tags:
        - Blog
      parameters:
        - name: blog_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the blog to update
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                cat_id:
                  type: integer
                  example: 2
                title:
                  type: string
                  example: "Updated Blog Post"
                slug:
                  type: string
                  example: "updated-blog-post"
                description1:
                  type: string
                  example: "Updated content of the blog."
                description2:
                  type: string
                  example: "Updated additional content of the blog."
                update_date:
                  type: string
                  format: date
                  example: "2023-04-22"
                image1:
                  type: string
                  format: binary
      responses:
        '200':
          description: Blog updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Blog updated successfully"
        '400':
          description: Invalid input data
        '404':
          description: Blog not found
        '500':
          description: Internal server error
    
    delete:
      summary: Delete a blog
      operationId: deleteBlog
      tags:
        - Blog
      parameters:
        - name: blog_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the blog to delete
      responses:
        '200':
          description: Blog deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Blog deleted"
        '404':
          description: Blog not found
        '500':
          description: Internal server error
...